requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.0.0
chardet==5.2.0
colorama==0.4.6
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-bootstrap4==22.3
django-cleanup==8.0.0
django-crispy-forms==2.0
django-filter==23.2
django-import-export==3.2.0
django-ipware==5.0.0
django-social-share==2.3.0
django-summernote==0.8.20.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
numpy==1.21.6
odfpy==1.4.1
openpyxl==3.1.2
pandas==1.3.5
Pillow==9.3.0
psycopg2==2.9.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.6
webencodings==0.5.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
django-recaptcha